refact: migrar la logica de carga de frases a la pagina home
