+++ frontend/src/app/layout.tsx
@@ -1,6 +1,5 @@
 import '../styles/globals.css';
 import { inter } from '@/styles/fonts';
-import Navbar from '@/ui/navbar';
 
 export default function RootLayout({
   children,
@@ -12,7 +11,6 @@ export default function RootLayout({
       <body className={`${inter.className} antialised`}>
         <div className="container h-screen lg:w-1/3 flex mx-auto md:border border-gray-300 rounded-t-3xl">
           {children}
-          <Navbar />
         </div>
       </body>
     </html>
